requirements.txt
fastapi
uvicorn 
grpcio 
grpcio-tools
requests
